requirements.txt
groq
python-dotenv
tqdm
